requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
ollama
